requirements.txt
# --- FastAPI / ASGI ---
fastapi==0.110.0
uvicorn[standard]==0.29.0

# --- Azure Authentication ---
azure-identity==1.17.1
azure-core==1.30.1

# --- Azure OpenAI ---
openai==1.40.0

# --- Cognitive Search (REST only, no SDK needed) ---
requests==2.31.0

# --- Tokenization for chunking ---
tiktoken==0.7.0

# --- File ingestion / parsing ---
pypdf==4.0.1              # PDF
python-docx==1.1.0        # DOCX
beautifulsoup4==4.12.3    # HTML / XML
lxml==5.2.1               # Faster XML/HTML parsing
markdown==3.6             # MD parsing if needed
python-magic==0.4.27      # Optional: detect file types

# --- Utilities ---
pyyaml==6.0.1             # YAML for job definitions
python-dotenv==1.0.1      # Local dev env vars
numpy==1.26.4             # For vector ops
tenacity==8.3.0           # Retry if needed

# --- Logging ---
rich==13.7.0

# --- Optional (but useful for ingestion robustness) ---
chardet==5.2.0            # Encoding detection
